ref(project-creation): Extract feature selection into a sibling component
Pull the feature-selection cards out of ScmPlatformFeaturesCore into a new
ScmFeatureSelectionPanel rendered as a sibling rather than nested. Core is
now platform selection only: detected platform cards plus the manual picker.

The panel re-derives the resolved platform and feature availability from the
same lifted props and the deduped platform-detection and feature-meta queries,
so no shared state is introduced. Both hosts render the two side by side at
the same level, so onboarding and project creation are unchanged. Platform
selected and step viewed analytics stay in Core; the feature toggled analytic
moves to the panel.

Refs VDY-77
