+++ static/app/views/onboarding/components/scmPlatformFeaturesCore.tsx
@@ -1,4 +1,4 @@
-import {Fragment, useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {motion} from 'framer-motion';
 import {PlatformIcon} from 'platformicons';
 
@@ -11,13 +11,8 @@ import {Text} from '@sentry/scraps/text';
 import {closeModal, openConsoleModal} from 'sentry/actionCreators/modal';
 import {LoadingIndicator} from 'sentry/components/loadingIndicator';
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  getDisabledProducts,
-  platformProductAvailability,
-} from 'sentry/components/onboarding/productSelection';
-import {PLATFORM_PRODUCT_INFO} from 'sentry/data/platformProductInfo.generated';
-import {IconBroadcast, IconBusiness} from 'sentry/icons';
-import {t, tct} from 'sentry/locale';
+import {IconBroadcast} from 'sentry/icons';
+import {t} from 'sentry/locale';
 import type {Repository} from 'sentry/types/integrations';
 import type {OnboardingSelectedSDK} from 'sentry/types/onboarding';
 import type {PlatformKey} from 'sentry/types/platform';
@@ -26,13 +21,9 @@ import {isDisabledGamingPlatform} from 'sentry/utils/platform';
 import {useOrganization} from 'sentry/utils/useOrganization';
 
 import type {ScmAnalyticsFlow} from './scmAnalyticsFlow';
-import {ScmFeatureInfoCards} from './scmFeatureInfoCards';
-import {ScmFeatureSelectionCards} from './scmFeatureSelectionCards';
 import {ScmPlatformCard} from './scmPlatformCard';
 import {
-  FEATURE_DISPLAY_ORDER,
   getPlatformInfo,
-  getPlatformName,
   platformOptions,
   type ResolvedPlatform,
   shouldSuggestFramework,
@@ -40,7 +31,6 @@ import {
 } from './scmPlatformHelpers';
 import {ScmSearchControl} from './scmSearchControl';
 import {ScmVirtualizedMenuList} from './scmVirtualizedMenuList';
-import {useScmFeatureMeta} from './useScmFeatureMeta';
 import {useScmPlatformDetection} from './useScmPlatformDetection';
 
 const STEP_VIEWED_EVENT = {
@@ -50,10 +40,6 @@ const STEP_VIEWED_EVENT = {
 const PLATFORM_SELECTED_EVENT = {
   onboarding: 'onboarding.scm_platform_selected',
   'project-creation': 'project_creation.scm_platform_selected',
-} as const;
-const FEATURE_TOGGLED_EVENT = {
-  onboarding: 'onboarding.scm_platform_feature_toggled',
-  'project-creation': 'project_creation.scm_platform_feature_toggled',
 } as const;
 const CHANGE_PLATFORM_CLICKED_EVENT = {
   onboarding: 'onboarding.scm_platform_change_platform_clicked',
@@ -69,19 +55,18 @@ interface ScmPlatformFeaturesCoreProps {
   onClearProjectDetailsForm: () => void;
   onFeaturesChange: (features: ProductSolution[] | undefined) => void;
   onPlatformChange: (platform: OnboardingSelectedSDK | undefined) => void;
-  selectedFeatures: ProductSolution[] | undefined;
   selectedPlatform: OnboardingSelectedSDK | undefined;
   selectedRepository: Repository | undefined;
 }
 
 /**
- * Core platform-and-features selection slice shared by the SCM onboarding
- * step (`ScmPlatformFeatures`) and the SCM-first project creation surface.
- * Renders the auto-detected platform cards (when an SCM repo is connected
- * and platforms were detected), the manual platform search dropdown, and
- * the feature selection / info cards. Owns platform detection, feature
- * metadata fetching, manual-picker toggle state, and the platform-selected
- * / feature-toggled / step-viewed / change-platform analytics.
+ * Platform-selection slice shared by the SCM onboarding step
+ * (`ScmPlatformFeatures`) and the SCM-first project creation surface. Renders
+ * the auto-detected platform cards (when an SCM repo is connected and platforms
+ * were detected) and the manual platform search dropdown. Feature selection is
+ * rendered separately as a sibling (`ScmFeatureSelectionPanel`). Owns platform
+ * detection, manual-picker toggle state, and the platform-selected /
+ * step-viewed / change-platform analytics.
  *
  * Does NOT render the step's surrounding chrome (page heading, "Choose
  * your SDK" subheading and description, Back / Continue footer). Hosts
@@ -92,21 +77,11 @@ export function ScmPlatformFeaturesCore({
   onClearProjectDetailsForm,
   onFeaturesChange,
   onPlatformChange,
-  selectedFeatures,
   selectedPlatform,
   selectedRepository,
 }: ScmPlatformFeaturesCoreProps) {
   const {openModal} = useModal();
   const organization = useOrganization();
-  // Trial/billing framing (the "unlimited volume" banner and per-feature volume
-  // limits) only makes sense during new-org onboarding, where a fresh trial is
-  // always active. In SCM-first project creation the viewer is an existing org
-  // on an unknown plan, so we hide that framing rather than show numbers that
-  // may not apply.
-  const isOnboarding = analyticsFlow === 'onboarding';
-  // Fetch feature meta at step entry so billing-config is in flight (or cached)
-  // before the user reaches the feature cards below.
-  const {meta: featureMeta, isLoading: isFeatureMetaLoading} = useScmFeatureMeta();
 
   const [showManualPicker, setShowManualPicker] = useState(false);
   // Guards the auto-detect analytics event below so it fires once per repo.
@@ -144,11 +119,6 @@ export function ScmPlatformFeaturesCore({
     isError: isDetectionError,
   } = useScmPlatformDetection(selectedRepository);
 
-  const currentFeatures = useMemo(
-    () => selectedFeatures ?? [ProductSolution.ERROR_MONITORING],
-    [selectedFeatures]
-  );
-
   const resolvedPlatforms = useMemo(
     () =>
       detectedPlatforms.reduce<ResolvedPlatform[]>((acc, detected) => {
@@ -164,8 +134,6 @@ export function ScmPlatformFeaturesCore({
   const detectedPlatformKey = resolvedPlatforms[0]?.platform;
   // Derive platform from explicit selection, falling back to first detected
   const currentPlatformKey = selectedPlatform?.key ?? detectedPlatformKey;
-
-  const currentPlatformName = getPlatformName(currentPlatformKey);
 
   // Adopt the first detected platform once per repo when the user hasn't
   // explicitly chosen one: commit it to the host so flows without a Continue
@@ -195,90 +163,6 @@ export function ScmPlatformFeaturesCore({
     analyticsFlow,
     setPlatform,
   ]);
-
-  // Wizard-driven platforms render an informational variant since the wizard CLI
-  // owns product configuration and toggles aren't actionable.
-  const featureMode = useMemo<'toggleable' | 'informational' | 'none'>(() => {
-    if (!currentPlatformKey) {
-      return 'none';
-    }
-    if (currentPlatformKey in platformProductAvailability) {
-      return 'toggleable';
-    }
-    if (currentPlatformKey in PLATFORM_PRODUCT_INFO) {
-      return 'informational';
-    }
-    return 'none';
-  }, [currentPlatformKey]);
-
-  const availableFeatures = useMemo(() => {
-    if (!currentPlatformKey || featureMode === 'none') {
-      return [];
-    }
-    const sourceProducts =
-      featureMode === 'toggleable'
-        ? platformProductAvailability[currentPlatformKey]
-        : PLATFORM_PRODUCT_INFO[currentPlatformKey];
-    const features = new Set<ProductSolution>([
-      ProductSolution.ERROR_MONITORING,
-      ...(sourceProducts ?? []),
-    ]);
-    return FEATURE_DISPLAY_ORDER.filter(f => features.has(f));
-  }, [currentPlatformKey, featureMode]);
-
-  const disabledProducts = useMemo(
-    () => getDisabledProducts(organization),
-    [organization]
-  );
-
-  const handleToggleFeature = useCallback(
-    (feature: ProductSolution) => {
-      if (disabledProducts[feature]) {
-        disabledProducts[feature]?.onClick?.();
-        return;
-      }
-
-      const wasEnabled = currentFeatures.includes(feature);
-      const newFeatures = new Set(
-        wasEnabled
-          ? currentFeatures.filter(f => f !== feature)
-          : [...currentFeatures, feature]
-      );
-
-      // Profiling requires tracing — mirror the constraint from ProductSelection
-      if (availableFeatures.includes(ProductSolution.PROFILING)) {
-        if (
-          feature === ProductSolution.PROFILING &&
-          newFeatures.has(ProductSolution.PROFILING)
-        ) {
-          newFeatures.add(ProductSolution.PERFORMANCE_MONITORING);
-        } else if (
-          feature === ProductSolution.PERFORMANCE_MONITORING &&
-          !newFeatures.has(ProductSolution.PERFORMANCE_MONITORING)
-        ) {
-          newFeatures.delete(ProductSolution.PROFILING);
-        }
-      }
-
-      onFeaturesChange(Array.from(newFeatures));
-
-      trackAnalytics(FEATURE_TOGGLED_EVENT[analyticsFlow], {
-        organization,
-        feature,
-        enabled: !wasEnabled,
-        platform: currentPlatformKey ?? '',
-      });
-    },
-    [
-      currentFeatures,
-      onFeaturesChange,
-      disabledProducts,
-      availableFeatures,
-      organization,
-      currentPlatformKey,
-      analyticsFlow,
-    ]
-  );
 
   const applyPlatformSelection = (sdk: OnboardingSelectedSDK) => {
     onPlatformChange(sdk);
@@ -428,147 +312,94 @@ export function ScmPlatformFeaturesCore({
     hasDetectedPlatforms &&
     (!currentPlatformKey || currentPlatformIsDetected);
 
-  return (
-    <Fragment>
-      {showDetectedPlatforms ? (
-        <MotionStack
-          key="detected"
-          initial={{opacity: 0}}
-          animate={{opacity: 1}}
-          gap="md"
-          width="100%"
-        >
-          <Flex justify="between" align="center">
-            <Flex align="center" gap="sm">
-              <IconBroadcast size="sm" />
-              <Text bold size="md" density="comfortable">
-                {t('Auto-detected from your repository')}
-              </Text>
-            </Flex>
-            <Button size="xs" variant="link" onClick={handleChangePlatformClick}>
-              {isDetecting
-                ? t('Skip detection and select manually')
-                : t("Doesn't look right? Change platform")}
-            </Button>
+  return showDetectedPlatforms ? (
+    <MotionStack
+      key="detected"
+      initial={{opacity: 0}}
+      animate={{opacity: 1}}
+      gap="md"
+      width="100%"
+    >
+      <Flex justify="between" align="center">
+        <Flex align="center" gap="sm">
+          <IconBroadcast size="sm" />
+          <Text bold size="md" density="comfortable">
+            {t('Auto-detected from your repository')}
+          </Text>
+        </Flex>
+        <Button size="xs" variant="link" onClick={handleChangePlatformClick}>
+          {isDetecting
+            ? t('Skip detection and select manually')
+            : t("Doesn't look right? Change platform")}
+        </Button>
+      </Flex>
+      <Stack gap="lg" width="100%">
+        {isDetecting ? (
+          <Flex justify="center">
+            <LoadingIndicator mini />
           </Flex>
-          <Stack gap="lg" width="100%">
-            {isDetecting ? (
-              <Flex justify="center">
-                <LoadingIndicator mini />
-              </Flex>
-            ) : (
-              <Grid
-                columns={{
-                  xs: '1fr',
-                  md: `repeat(${resolvedPlatforms.length}, minmax(200px, 1fr))`,
-                }}
-                width="100%"
-                justify="center"
-                gap="md"
-                role="radiogroup"
-              >
-                {resolvedPlatforms.map(({platform, info}) => (
-                  <ScmPlatformCard
-                    key={platform}
-                    platform={platform}
-                    name={info.name}
-                    type={info.type}
-                    isSelected={currentPlatformKey === platform}
-                    onClick={() => handleSelectDetectedPlatform(platform)}
-                  />
-                ))}
-              </Grid>
-            )}
-          </Stack>
-        </MotionStack>
-      ) : (
-        <MotionStack
-          key="manual"
-          gap="md"
-          width="100%"
-          initial={{opacity: 0}}
-          animate={{opacity: 1}}
-        >
-          <Flex justify="between" align="center">
-            <Flex align="center" gap="sm">
-              <Text bold size="md" density="comfortable">
-                {t('Select a platform')}
-              </Text>
-            </Flex>
-            {hasScmConnected && !isDetectionError && hasDetectedPlatforms && (
-              <Button size="xs" variant="link" onClick={handleBackToRecommended}>
-                {t('Back to recommended platforms')}
-              </Button>
-            )}
-          </Flex>
-          <Select<(typeof platformOptions)[number]>
-            placeholder={t('Search SDKs...')}
-            options={manualPickerOptions}
-            value={currentPlatformKey ?? null}
-            onChange={option => {
-              if (option) {
-                handleManualPlatformSelect(option);
-              }
+        ) : (
+          <Grid
+            columns={{
+              xs: '1fr',
+              md: `repeat(${resolvedPlatforms.length}, minmax(200px, 1fr))`,
             }}
-            searchable
-            components={{
-              Control: ScmSearchControl,
-              MenuList: ScmVirtualizedMenuList,
-            }}
-            styles={{container: base => ({...base, width: '100%'})}}
-          />
-        </MotionStack>
-      )}
-      {featureMode !== 'none' && (
-        <MotionStack layout="position" width="100%">
-          <Stack gap="2xl" paddingTop="xs">
-            {isOnboarding && (
-              <Flex
-                padding="lg"
-                background="secondary"
-                border="secondary"
-                radius="md"
-                gap="lg"
-              >
-                <IconBusiness size="lg" variant="accent" />
-                <Text size="md" density="comfortable">
-                  {tct(
-                    'You’ve got [bold:unlimited volume for 14 days] to try out everything. After that, free plan volumes apply ⋅ No credit card required',
-                    {
-                      bold: (
-                        <Text as="span" bold variant="accent">
-                          {null}
-                        </Text>
-                      ),
-                    }
-                  )}
-                </Text>
-              </Flex>
-            )}
-            {featureMode === 'toggleable' ? (
-              <ScmFeatureSelectionCards
-                availableFeatures={availableFeatures}
-                selectedFeatures={currentFeatures}
-                disabledProducts={disabledProducts}
-                onToggleFeature={handleToggleFeature}
-                featureMeta={featureMeta}
-                isVolumeLoading={isFeatureMetaLoading}
-                showVolume={isOnboarding}
+            width="100%"
+            justify="center"
+            gap="md"
+            role="radiogroup"
+          >
+            {resolvedPlatforms.map(({platform, info}) => (
+              <ScmPlatformCard
+                key={platform}
+                platform={platform}
+                name={info.name}
+                type={info.type}
+                isSelected={currentPlatformKey === platform}
+                onClick={() => handleSelectDetectedPlatform(platform)}
               />
-            ) : (
-              <ScmFeatureInfoCards
-                availableFeatures={availableFeatures}
-                disabledProducts={disabledProducts}
-                featureMeta={featureMeta}
-                platformName={currentPlatformName}
-                isVolumeLoading={isFeatureMetaLoading}
-                showVolume={isOnboarding}
-              />
-            )}
-          </Stack>
-        </MotionStack>
-      )}
-    </Fragment>
+            ))}
+          </Grid>
+        )}
+      </Stack>
+    </MotionStack>
+  ) : (
+    <MotionStack
+      key="manual"
+      gap="md"
+      width="100%"
+      initial={{opacity: 0}}
+      animate={{opacity: 1}}
+    >
+      <Flex justify="between" align="center">
+        <Flex align="center" gap="sm">
+          <Text bold size="md" density="comfortable">
+            {t('Select a platform')}
+          </Text>
+        </Flex>
+        {hasScmConnected && !isDetectionError && hasDetectedPlatforms && (
+          <Button size="xs" variant="link" onClick={handleBackToRecommended}>
+            {t('Back to recommended platforms')}
+          </Button>
+        )}
+      </Flex>
+      <Select<(typeof platformOptions)[number]>
+        placeholder={t('Search SDKs...')}
+        options={manualPickerOptions}
+        value={currentPlatformKey ?? null}
+        onChange={option => {
+          if (option) {
+            handleManualPlatformSelect(option);
+          }
+        }}
+        searchable
+        components={{
+          Control: ScmSearchControl,
+          MenuList: ScmVirtualizedMenuList,
+        }}
+        styles={{container: base => ({...base, width: '100%'})}}
+      />
+    </MotionStack>
   );
 }
